use-cases/01-agentic-ai-platform/requirements.txt
anthropic>=0.84.0
fastapi>=0.135.0
uvicorn>=0.41.0
pydantic>=2.12.0
redis>=7.2.0
aiohttp>=3.13.0
python-dotenv>=1.2.0
httpx>=0.28.0
